projects/attestio/tools/carousel-generator/requirements.txt
requests>=2.28.0
Pillow>=10.0.0
fpdf2>=2.7.0
